add choice back if package not found, check .gopac not found after enter CTL+C

+++ src/utils/detectKeyboard.ts
@@ -1,10 +1,18 @@
-import { deleteData } from './fileSystem'
+/**
+ * Gopack CLI
+ * @author Copyright(c) 2021 by Restu wahyu saputra
+ * MIT Licensed
+ */
+
+import { deleteData, existData } from './fileSystem'
 
 export const detectKeyboard = (): void => {
 	process.openStdin().on('keypress', function (_, key: Record<string, any>) {
-		if (key && key.name === 'c' && key.ctrl) {
-			deleteData()
-			process.exit()
+		if (existData() !== false) {
+			if (key && key.name === 'c' && key.ctrl) {
+				deleteData()
+				process.exit()
+			}
 		}
 	})
 }
